Added CRUD for movies, uploaded movie image

+++ routes/users.js
@@ -1,7 +1,8 @@
 const express = require('express');
 const router = express.Router();
 const { body, param, validationResult } = require('express-validator');
-const validateRequest = require('../middlewares/validation');
+const validateRequest = require('../middlewares/validation').validateRequest;
+const validateToken = require('../middlewares/validation').validateToken;
 const UserController = require('../controllers/userController');
 
 
@@ -11,6 +12,8 @@ const UserController = require('../controllers/userController');
  *   get:
  *     summary: Get all users
  *     description: Retrieve a list of all users.
+ *     security: 
+ *       - JWTAuth: []
  *     responses:
  *       200:
  *         description: A list of users.
@@ -23,11 +26,15 @@ const UserController = require('../controllers/userController');
  *       500:
  *         description: Internal server error.
  */
-router.get('/', async (req, res, next) => {
+router.get('/', 
+validateToken,
+async (req, res, next) => {
   try {
+    if(req.user.email != "admin@gmail.com") {
+      return res.status(401).json({ message: 'User not authorized' });
+    }
     const users = await UserController.getAllUsers(req, res, next);
-    console.log(users);
-    res.status(200).json(users);
+    return res.status(200).json(users);
   } catch (error) {
     next(error); // Pass the error to the error handling middleware
   }
@@ -39,6 +46,8 @@ router.get('/', async (req, res, next) => {
  *   get:
  *     summary: Get a specific user
  *     description: Retrieve details of a specific user by their ID.
+ *     security: 
+ *       - JWTAuth: []
  *     parameters:
  *       - in: path
  *         name: userID
@@ -60,9 +69,13 @@ router.get('/', async (req, res, next) => {
  */
 router.get('/:userID',
   param('userID').isInt().withMessage('User ID must be an integer'),
-//  validateRequest,
+  validateRequest,
+  validateToken,
   async (req, res, next) => {
     try {
+      if(req.user.userID != req.params.userID && req.user.email != "admin@gmail.com") {
+        return res.status(401).json({ message: 'User not authorized' });
+      }
       const user = await UserController.getUserById(req, res, next);
       if (!user) {
         return res.status(404).json({ message: 'User not found' });
@@ -80,6 +93,8 @@ router.get('/:userID',
  *   post:
  *     summary: Create a new user
  *     description: Create a new user with the provided details.
+ *     security: 
+ *       - JWTAuth: []
  *     requestBody:
  *       required: true
  *       content:
@@ -99,6 +114,7 @@ router.post('/',
   body('birthDate').notEmpty().withMessage('Birth date is required'),
   body('birthDate').isISO8601().withMessage('Invalid birth date format. Use YYYY-MM-DD format.'),
   body('birthDate').isAfter('1900-01-01').withMessage('Invalid birth date.'),
+  body('password').notEmpty().withMessage('Password is required'),
   validateRequest, //custom validation middleware
   async (req, res, next) => {
     try {
@@ -116,6 +132,8 @@ router.post('/',
  *   delete:
  *     summary: Delete a user
  *     description: Delete a specific user by their ID.
+ *     security: 
+ *       - JWTAuth: []
  *     parameters:
  *       - in: path
  *         name: userID
@@ -132,8 +150,12 @@ router.post('/',
 router.delete('/:userID',
   param('userID').isInt().withMessage('User ID must be an integer'),
   validateRequest,
+  validateToken,
   async (req, res, next) => {
     try {
+      if(req.user.userID != req.params.userID && req.user.email != "admin@gmail.com") {
+        return res.status(401).json({ message: 'User not authorized' });
+      }
       const user = await UserController.deleteUser(req, res, next);
       if (!user) {
         return res.status(404).json({ message: 'User not found' });
@@ -151,6 +173,8 @@ router.delete('/:userID',
  *   put:
  *     summary: Update a user
  *     description: Update the details of a specific user by their ID.
+ *     security: 
+ *       - JWTAuth: []
  *     parameters:
  *       - in: path
  *         name: userID
@@ -172,12 +196,16 @@ router.delete('/:userID',
  */
 router.put('/:userID',
   param('userID').isInt().withMessage('User ID must be an integer'),
-  body('email').isEmail().withMessage('Invalid email address'),
+  body('email').optional({ checkFalsy: true }).isEmail().withMessage('Invalid email address'),
   body('birthDate').optional({ checkFalsy: true }).isISO8601().withMessage('Invalid birth date format. Use YYYY-MM-DD format.'),
   body('birthDate').optional({ checkFalsy: true }).isAfter('1900-01-01').withMessage('Invalid birth date.'),
   validateRequest,
+  validateToken,
   async (req, res, next) => {
     try {
+      if(req.user.userID != req.params.userID && req.user.email != "admin@gmail.com") {
+        return res.status(401).json({ message: 'User not authorized' });
+      }
       const user = await UserController.updateUser(req, res, next);
       if (!user) {
         return res.status(404).json({ message: 'User not found' });
